+++ Pandora/Paradigm/Structure/Stack.hs
@@ -37,7 +37,7 @@ import Pandora.Paradigm.Structure.Ability.Nonempty (Nonempty)
 import Pandora.Paradigm.Structure.Ability.Nullable (Nullable (null))
 import Pandora.Paradigm.Structure.Ability.Zipper (Zipper)
 import Pandora.Paradigm.Structure.Ability.Focusable (Focusable (Focusing, focusing), Location (Head), focus)
-import Pandora.Paradigm.Structure.Ability.Deletable (Deletable (delete))
+import Pandora.Paradigm.Structure.Ability.Deletable (Deletable ((-=)))
 import Pandora.Paradigm.Structure.Ability.Insertable (Insertable ((+=)))
 import Pandora.Paradigm.Structure.Ability.Measurable (Measurable (Measural, measurement), Scale (Length), measure)
 import Pandora.Paradigm.Structure.Ability.Monotonic (Monotonic (reduce))
@@ -81,9 +81,9 @@ instance Substructure Tail Stack where
 	substructure (run . extract . run -> Nothing) = Store $ empty :*: lift . identity
 
 instance Deletable Stack where
-	delete _ (TU Nothing) = TU Nothing
-	delete x (TU (Just (Construct y ys))) = x == y ? TU ys
-		$ lift . Construct y . run . delete @Stack x $ TU ys
+	_ -= TU Nothing = TU Nothing
+	x -= TU (Just (Construct y ys)) = x == y ? TU ys
+		$ lift . Construct y . run . (-=) @Stack x $ TU ys
 
 filter :: forall a . Predicate a -> Stack a -> Stack a
 filter (Predicate p) = TU . extract