+++ CHANGELOG.md
@@ -346,3 +346,4 @@
 
 # 0.3.7
 * Rename `insert` method to `+=` in `Insertable` typeclass
+* Rename `delete` method to `-=` in `Deletable` typeclass